+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] / 2024-02-17
+### Features
+- Support net core plugin.
+### Added
+- Add `CodeAnalysisCodeDom` to support net code.
+- Update `CodeAnalysisCodeDomService` to work with multiple source code.
+### Tests
+- Test `GistGithubUtils` download string.
+- Add `Newtonsoft.Json` to support `GistGithubUtils` download string.
+
 ## [1.0.6] / 2024-01-27
 ### Features
 - Using `ricaun.Revit.UI.Tasks`
@@ -49,6 +59,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [x] AutoUpdater
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.1.0]: ../../compare/1.0.6...1.1.0
 [1.0.6]: ../../compare/1.0.5...1.0.6
 [1.0.5]: ../../compare/1.0.4...1.0.5
 [1.0.4]: ../../compare/1.0.3...1.0.4